+++ database/migrations/2020_05_10_100000_create_translations_table.php
@@ -18,11 +18,10 @@ class CreateTranslationsTable extends Migration
             $table->string('translatable_attribute');
             $table->text('value');
             $table->string('locale', 24)->nullable()->comment('RFC 5646. See: http://www.rfc-editor.org/rfc/rfc5646.txt');
-            $table->boolean('is_archived')->default(false);
             $table->timestamps();
 
             // Indices
-            $table->index(['translatable_type', 'translatable_id', 'locale'], 'translatable-index');
+            $table->index(['translatable_type', 'translatable_id', 'locale'], 'translation');
         });
     }