+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Strategy to store translations in the separate table for specific model.
+
+### Removed
+- `translate` method on the model.
 
 ## [0.7.4] - 2021-11-29
 ### Changed